requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
plotly
python-dotenv